CHROMIUM: arm64: dts: rockchip: change trigger mode for trackpad on nefario
As trackpad reported the invalid data 0xff when keep pressing it.
That indicate the hardware hasn't been triggered the interrupt of
trackpad, but the system need get the report data from trackpad.

The trigger mode of edge falling maybe trigger the interrupt, but the
trackpad's firmware don't think so. Change trigger mode for the Elan's
trackpad.

BUG=b:63537905
TEST=Tested on Nefario.

Change-Id: I3015e274640d67083c6f2c2da6b8798df494e07b
Signed-off-by: Caesar Wang <wxt@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/612933
Reviewed-by: Douglas Anderson <dianders@chromium.org>