CHROMIUM: arm64: dts: rockchip: change trigger mode for elan on gru
In http://crosreview.com/522266 we changed the trigger mode for the
elan trackpads and touchscreens on rk3399-gru-bob.  However, we didn't
change it for gru.  That meant that when we created rk3399-gru-nefario
we accidentally copied the wrong example, leading to the need to fix
it as per http://crosreview.com/612933

Let's fix it in gru.

BUG=None
TEST=Ideally, check touchscreen/trackpad on gru.

Change-Id: I2c6f96a26bba566bea2de5e57fc2b8d6ba1095ac
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/613682
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>