UPSTREAM: media: cec-pin: fix irq handling
The free_irq() function could be called from interrupt context,
which is invalid. Move this to the thread.

In the interrupt handler we just request that the thread disables
the irq. This is done through an atomic so we don't need to add
any spinlocks.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>
(cherry picked from commit cb7474949371ba8d7591a62924f05b627a2601d9)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Change-Id: If4d4fc4fa01e053206fd45d40555e3aa3c72bd10
Reviewed-on: https://chromium-review.googlesource.com/1021356
Commit-Ready: Sean Paul <seanpaul@google.com>
Tested-by: Sean Paul <seanpaul@google.com>
Reviewed-by: Sean Paul <seanpaul@google.com>