CHROMIUM: Don't trust DPCD for panel earlier than eDP 1.4
Swanky panel has DPCD 702h = 0xf6 which means panel only support
AUX to adjust brightness not the PWM pin. But actually it is
opposite of what it is claiming in DPCD register.

This CL make driver to always use PWM pin to adjust brightness
for panel eariler than eDP 1.4 even when the panel DPCD said
that PWM pin is not supported.

BUG=chromium:755752
TEST=Swanky brightness settings work again.

Change-Id: I305b72fc9c7f19581cc350bd3475a18cacef1da8
Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/622047
Reviewed-by: Todd Broch <tbroch@chromium.org>