CHROMIUM: net: Increase net.unix.max_dgram_qlen.
The parameter is 10 by default and 600 on Android. This CL sets the
value to 60, large enough for ARC to pass the CTS tests that depend
on logcat and thus are largely affected by the config. Experiments on
a Reef device running the test noted in the TEST field showed:

   max_dgram_qlen  pass_rate
   -------------------------
        10            12%
        30            80%
        50            98%
        60           100%
       100           100%
       600           100%

Unfortunately, normal method like sysctl or /proc/sys cannot modify the
parameter inside namespaces. To quickly fix it up for the Android
container, we directly patch the kernel for now.

BUG=b:64655659
BUG=chromium:758081
TEST=test_that reef cheets_CTS_N.7.1_r8.x86.CtsServicesHostTestCases

Change-Id: Id00efd1a300620f57eb76b361d0b71c895dce2c2
Signed-off-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/627966
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>
(cherry picked from commit fa7f9d27f08874e13dcc6730a11985d897baf349)
Reviewed-on: https://chromium-review.googlesource.com/636923