CHROMIUM: kaiser: delete KAISER_REAL_SWITCH option
We fail to see what CONFIG_KAISER_REAL_SWITCH is for: it seems to be
left over from early development, and now just obscures tricky parts
of the code.  Delete it before adding PCIDs, or nokaiser boot option.

(Or if there is some good reason to keep the option, then it needs
a help text - and a "depends on KAISER", so that all those without
KAISER are not asked the question.)

BUG=chromium:781913
TEST=Build, deploy on caroline.

Originally-from: Dave Hansen <dave.hansen@intel.com>

Conflicts:
arch/x86/entry/entry_64.S (not in this tree)
arch/x86/kernel/entry_64.S (patched instead of that)

Change-Id: I94570de669a5721030a1ab994cb961fbfc93b4c1
Signed-off-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/757208
Reviewed-by: Guenter Roeck <groeck@chromium.org>