CHROMIUM: kaiser: tidied up kaiser_add/remove_mapping slightly
Yes, unmap_pud_range_nofree()'s declaration ought to be in a
header file really, but I'm not sure we want to use it anyway:
so for now just declare it inside kaiser_remove_mapping().
And there doesn't seem to be such a thing as unmap_p4d_range(),
even in a 5-level paging tree.

BUG=chromium:781913
TEST=Build, deploy on caroline.

Change-Id: I6cacea1f740dd80451dae1ba4fae21a13be3343d
Originally-from: Dave Hansen <dave.hansen@intel.com>
Signed-off-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/757203
Reviewed-by: Guenter Roeck <groeck@chromium.org>