UPSTREAM: drm/i915/psr: Move hsw_enable_source after enabling sink.
No functional change is expected here since at this point
PSR is not allowed to go to any active state. In other
words, not really enabled.

However let's do in a separated patch so it gets clear
on what is change and specially it can helps on bisect
case if we figure something has caused changes in behaviour.

But this needs to be done before we make the vfunc to
enable source to be in parity with VLV implementation.

Cc: Jani Nikula <jani.nikula@intel.com>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>
Cc: Vathsala Nagaraju <vathsala.nagaraju@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Reviewed-by: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20170907230041.22978-8-rodrigo.vivi@intel.com
(cherry picked from commit 196cebddde80dd06b2de557ea393082bf5a06c3d)
from v4.15-rc1
Signed-off-by: Tarun Vyas <tarun.vyas@intel.com>

BUG=b:70920211, b:109823283
TEST=No *Potential atomic update error* in dmesg
with external display connected and browser opened
and cursor blinking in address bar for PSR entry-exit
TEST=No other regressions on KBL/APL/BYT based devices

Change-Id: Ie990a694339d25d492859eb19ad7eb8f409800f1
Reviewed-on: https://chromium-review.googlesource.com/1135388
Commit-Ready: Nathan Ciobanu <nathan.d.ciobanu@intel.com>
Tested-by: Casey G Bowman <casey.g.bowman@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>