FROMLIST: ASoC: Intel: Skylake: Parse nhlt to populate clock information
NHLT is parsed to fill the SSP clock information. If an endpoint is present
for a SSP, then the corresponding SSP clocks are created.

MCLK is consistent across endpoints and configuration for an SSP. So NHLT
is queried for the first endpoint for an SSP and MCLK information for that
SSP is populated.

For SCLK/SCLKFS, the best fit is queried from the NHLT configurations which
matches the clock rate requested. Best fit is decided based on below:

1. If rate matches with multiple configurations, then the first
configuration is selected.

2. If for a selected fs and bits_per_sample, there are multiple endpoint
configuration match, then the configuration with max number of channels is
selected. So, the user has to set the rate which fits max number of
channels

Max of 18 clocks created, 3 for each SSP and there are 6 SSPs. There are
also 3 parent clocks XTAL, Cardinal and PLL. This patch populates the all
the clock information.

Cardinal/PLL are fixed for all platforms. For XTAL, APL/GLK uses 19.2MHz,
whereas for SKL/APL it is 24MHz.

Signed-off-by: Jaikrishna Nemallapudi <jaikrishnax.nemallapudi@intel.com>
Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>

Link:https://patchwork.kernel.org/patch/9955479
(cherry picked from commit 4a8b037a154480df084d872709438f7dbbceda3b
git://git.kernel.org/pub/scm/linux/kernel/git/vkoul/sound.git
origin/intel)

BUG=b:62106805
TEST=With entire series merged, amic capture should have valid data.
(NOTE:Test needs to be done without any silent playback in bg)

Change-Id: I3fd8bbe1c9e103795ed51689310e4e7b4b99c4df
Signed-off-by: Naveen M <naveen.m@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/672764
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Shruthi Sudhakar <shruthi.sudhakar@intel.com>