CHROMIUM: drm/img-rogue: disable run-time BVNC acquisition
Run-time BVNC acquisition works by powering on the device, reading a register
and then powering the device off again. However, it was found that, in this
particular scenario, powering off was adding ~1 second to the boot time.
Since there isn't currently a need for this feature (as only a single BVNC
is supported), disable it for the time being.

BUG=b:64429939
TEST=Boot to UI

Change-Id: I0fe845125f2f7984d1da36d513d052a10111c444
Signed-off-by: Frank Binns <frank.binns@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/673443
Tested-by: Lin Shu-Fa <afar.lin@imgtec.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>